// resources/manager/containers/app/frontend/poc-plataforma-app/src/app/funcionalidades/relatorios/models/proposta.model.ts
export class Proposta {
  id: string;
  idPessoa: string;
  numero: string;
  proposta: number;
  produto: string;
  dataContratacao: Date;
  cpfCnpj: string;
  canalContratacao: string;
  dataSituacao: Date;
  nomePessoa: string;
  dataInicio: Date;
  dataFim: Date;
  pa: string;
  situacao: string;
  valorPremio: string;
  valorPremioMensal: string;
  valorCapitalSegurado: string;
  dataVenda: Date;
  periodicidade: string;
  nomeVendedor: string;
  cpfVendedor: string;
  motivoCancelamento: string;
  dataCancelamento: Date;
  unidadeInst: string;
  idInstituicao: string;
  cooperativa: string;
}